+++ CHANGELOG.md
@@ -19,6 +19,11 @@ x.xx.x Release notes (yyyy-MM-dd)
 
 * `Int8` properties defined in Swift are now treated as integers, rather than
   booleans.
+* NSPredicates created using +predicateWithValue: are now supported.
+
+### Bugfixes
+
+* Compound AND predicates with no subpredicates now correctly match all objects.
 
 0.91.5 Release notes (2015-04-28)
 =============================================================