+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-x.xx.x Release notes (yyyy-MM-dd)
+0.92.0 Release notes (2015-05-05)
 =============================================================
 
 ### API breaking changes
@@ -19,7 +19,7 @@ x.xx.x Release notes (yyyy-MM-dd)
 
 * `Int8` properties defined in Swift are now treated as integers, rather than
   booleans.
-* NSPredicates created using +predicateWithValue: are now supported.
+* NSPredicates created using `+predicateWithValue:` are now supported.
 
 ### Bugfixes